+++ ProjectSetupV2/Models/Context/Tasks.cs
@@ -5,10 +5,10 @@ namespace ProjectSetupV2.Models.Context
 {
     public partial class Tasks
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Task { get; set; }
-        public long? JobId { get; set; }
-        public long? BusinessValuesId { get; set; }
+        public int? JobId { get; set; }
+        public int? BusinessValuesId { get; set; }
         public double? TasksRate { get; set; }
         public string Status { get; set; }
